Split shell agent transport auth tests
